Updated inscrire & authentication
