+++ src/app/authentification/authentification.component.html
@@ -1,28 +1,3 @@
-<div class="auth-wrapper">
-    <div class="auth-inner">
-      <form [formGroup]="authForm" (ngSubmit)="onSubmit()">
-        <h2>S'identifier</h2>
-  
-        <div class="form-group">
-          <label for="email">Email</label>
-          <input type="email" formControlName="email" class="form-control" placeholder="Enter email">
-          <div *ngIf="isInvalidAndTouched('email')" class="invalid-feedback">
-            Please enter a valid email address.
-          </div>
-        </div>
-  
-        <div class="form-group">
-          <label for="password">Password</label>
-          <input type="password" formControlName="password" class="form-control" placeholder="Enter password">
-          <div *ngIf="isInvalidAndTouched('password')" class="invalid-feedback">
-            Please enter a valid password.
-          </div>
-        </div>
-  
-        <div class="form-group">
-          <button type="submit" [disabled]="!authForm.valid" class="btn btn-primary btn-block">Submit</button>
-        </div>
-      </form>
-    </div>
-  </div>
-  
+<div>
+  <app-connecter></app-connecter>
+</div>
