+++ src/app/authentification/inscrire/company-inscription/company-inscription.component.html
@@ -1,0 +1,32 @@
+<form #companyForm="ngForm" (ngSubmit)="onCompanySubmit(companyForm.value)">
+    <div class="form-group">
+      <label for="name">Nom de l'entreprise</label>
+      <input type="text" class="form-control" id="name" required ngModel name="name">
+    </div>
+    <div class="form-group">
+      <label for="description">Description</label>
+      <textarea class="form-control" id="description" required ngModel name="description"></textarea>
+    </div>
+    <div class="form-group">
+      <label for="email">Email</label>
+      <input type="email" class="form-control" id="email" required ngModel name="email">
+    </div>
+    <div class="form-group">
+      <label for="password">Mot de passe</label>
+      <input type="password" class="form-control" id="password" required ngModel name="password">
+    </div>
+    <div class="form-group">
+      <label for="website">Site Web</label>
+      <input type="text" class="form-control" id="website" ngModel name="website">
+    </div>
+    <div class="form-group">
+      <label for="state">État</label>
+      <input type="text" class="form-control" id="state" required ngModel name="state">
+    </div>
+    <div class="form-group">
+      <label for="country">Pays</label>
+      <input type="text" class="form-control" id="country" required ngModel name="country">
+    </div>
+    <button type="submit" class="btn btn-primary" [disabled]="!companyForm.valid">S'inscrire</button>
+  </form>
+
