+++ src/app/authentification/inscrire/inscrire.component.ts
@@ -6,16 +6,13 @@ import { Component } from '@angular/core';
   styleUrls: ['./inscrire.component.css']
 })
 export class InscrireComponent {
-  inscriptionType = 'user'; // 'user' ou 'company'
+  inscriptionType = 'user';
 
   setInscriptionType(type: string) {
     this.inscriptionType = type;
   }
 
-
   onCompanySubmit(formValue: any) {
-    // Utilisez le service pour envoyer les données du formulaire à votre backend
     console.log(formValue);
-    // Exemple : this.myService.createCompany(formValue).subscribe(...);
   }
 }
