--- src/app/inscrire/inscrire.component.html
@@ -1,22 +1,0 @@
-<div class="reg-wrapper">
-  <div class="reg-inner">
-    <h2>S'inscrire</h2>
-
-    <div class="nav-inscription">
-      <div class="nav-buttons">
-        <button class="button-tab" [class.active]="inscriptionType === 'user'" (click)="setInscriptionType('user')">User</button>
-        <button class="button-tab" [class.active]="inscriptionType === 'company'" (click)="setInscriptionType('company')">Company</button>
-      </div>
-    </div>
-
-    <!-- Conteneur du formulaire d'inscription pour l'utilisateur -->
-    <div class="form-container" *ngIf="inscriptionType === 'user'">
-      <app-user-inscription></app-user-inscription>
-    </div>
-
-    <!-- Conteneur du formulaire d'inscription pour l'entreprise -->
-    <div class="form-container" *ngIf="inscriptionType === 'company'">
-      <app-company-inscription></app-company-inscription>
-    </div>
-  </div>
-</div>
