--- src/app/inscrire/user-inscription/user-inscription.component.html
@@ -1,36 +1,0 @@
-<form #userForm="ngForm" (ngSubmit)="onUserSubmit(userForm.value)">
-  <div class="form-group">
-    <label for="firstname">Prénom</label>
-    <input type="text" class="form-control" id="firstname" required ngModel name="firstname">
-  </div>
-  <div class="form-group">
-    <label for="lastname">Nom de famille</label>
-    <input type="text" class="form-control" id="lastname" required ngModel name="lastname" minlength="3">
-  </div>
-  <div class="form-group">
-    <label for="phoneNumber">Numéro de téléphone</label>
-    <input type="text" class="form-control" id="phoneNumber" required ngModel name="phoneNumber" minlength="8" maxlength="50">
-  </div>
-  <div class="form-group">
-    <label for="email">Email</label>
-    <input type="email" class="form-control" id="email" required ngModel name="email">
-  </div>
-  <div class="form-group">
-    <label for="password">Mot de passe</label>
-    <input type="password" class="form-control" id="password" required ngModel name="password">
-  </div>
-  <div class="form-group">
-    <label for="address">Adresse</label>
-    <input type="text" class="form-control" id="address" required ngModel name="address" minlength="3" maxlength="32">
-  </div>
-  <div class="form-group">
-    <label for="gender">Genre</label>
-    <select class="form-control" id="gender" required ngModel name="gender">
-      <option value="male">Homme</option>
-      <option value="female">Femme</option>
-      <option value="other">Autre</option>
-    </select>
-  </div>
-  <!-- Ajoutez d'autres champs si nécessaire -->
-  <button type="submit" class="btn btn-primary" [disabled]="!userForm.valid">S'inscrire</button>
-</form>
